Added self-logging
Self-logging is the logging of any actions and particularly of failures of
the program itself. This is separate from forwarding the HMC log entries.

Self-logging can be configured with its destination (stdout, stderr),
log message format and time format therein.

At this point, the configuration can only be done in the config file,
and not in command line options.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>