Added support for multiple forwardings
This change adds support for defining multiple log forwardings
in the configuration. A log forwarding is mainly a set of logs
to collect, and a destination to forward them to. This allows
for example to send some logs to QRadar and some (same or others)
to logDNA.

This required introducing support for hierarchical config parameters,
which was done by redesigning the whole config parameter processing
to use JSON schema for the config file structure. This also
caused the command line arguments that could be used to
override config file parameters, to go away (too complex).

Signed-off-by: Andreas Maier <maiera@de.ibm.com>